Implement ability to rerender photos using different API endpoints
